refactor: extract request dispatch router
