+++ src/global_state.rs
@@ -1,4 +1,5 @@
 mod diagnostics;
+mod dispatch;
 mod dispatcher;
 pub(crate) mod event_loop;
 mod handlers;
